Refactor: libcrmcommon: Make various schema stuff non-static.
This is the minimum amount necessary to make the function unit testable.
None of this is intended to ever become public.
